@@ -23,46 +23,44 @@
     "require": {
         "php": " >=8.3",
         "ext-json": "*",
-        "babdev/pagerfanta-bundle": "^2.1",
+        "babdev/pagerfanta-bundle": "^3.8",
         "ibexa/content-forms": "~5.0.x-dev",
         "ibexa/core": "~5.0.x-dev",
         "ibexa/design-engine": "~5.0.x-dev",
-        "ibexa/fieldtype-richtext": "~5.0.x-dev",
         "ibexa/polyfill-php82": "^1.0",
         "ibexa/rest": "~5.0.x-dev",
         "ibexa/search": "~5.0.x-dev",
         "ibexa/user": "~5.0.x-dev",
-        "jms/translation-bundle": "^1.5",
+        "jms/translation-bundle": "^2.4",
         "knplabs/knp-menu-bundle": "^3.0",
         "mck89/peast": "^1.9",
-        "symfony/asset": "^5.0",
-        "symfony/config": "^5.0",
-        "symfony/console": "^5.0",
-        "symfony/dependency-injection": "^5.0",
+        "symfony/asset": "^6.4",
+        "symfony/config": "^6.4",
+        "symfony/console": "^6.4",
+        "symfony/dependency-injection": "^6.4",
         "symfony/deprecation-contracts": "^2.5",
-        "symfony/event-dispatcher": "^5.0",
-        "symfony/filesystem": "^5.0",
-        "symfony/form": "^5.0",
-        "symfony/http-foundation": "^5.0",
-        "symfony/http-kernel": "^5.0",
-        "symfony/options-resolver": "^5.0",
-        "symfony/routing": "^5.0",
-        "symfony/security-core": "^5.0",
-        "symfony/security-http": "^5.0",
-        "symfony/translation": "^5.0",
-        "symfony/validator": "^5.0",
-        "symfony/webpack-encore-bundle": "^v1.17",
-        "symfony/yaml": "^5.0",
+        "symfony/event-dispatcher": "^6.4",
+        "symfony/filesystem": "^6.4",
+        "symfony/form": "^6.4",
+        "symfony/http-foundation": "^6.4",
+        "symfony/http-kernel": "^6.4",
+        "symfony/options-resolver": "^6.4",
+        "symfony/routing": "^6.4",
+        "symfony/security-core": "^6.4",
+        "symfony/security-http": "^6.4",
+        "symfony/translation": "^6.4",
+        "symfony/validator": "^6.4",
+        "symfony/webpack-encore-bundle": "^1.17",
+        "symfony/yaml": "^6.4",
         "twig/intl-extra": "^3.0",
         "twig/string-extra": "^3.0",
         "twig/twig": "^3.0",
-        "willdurand/js-translation-bundle": "^4.0"
+        "willdurand/js-translation-bundle": "^6.1"
     },
     "require-dev": {
         "ext-zip": "*",
         "dama/doctrine-test-bundle": "^v6.7",
         "ibexa/behat": "~5.0.x-dev",
-        "ibexa/ci-scripts": "^0.2@dev",
         "ibexa/code-style": "~2.0.0",
         "ibexa/doctrine-schema": "~5.0.x-dev",
         "ibexa/http-cache": "~5.0.x-dev",
