@@ -9,20 +9,20 @@ declare(strict_types=1);
 namespace Ibexa\AdminUi\Form\Type\ChoiceList\Loader;
 
 use Ibexa\Contracts\Core\Repository\Values\Content\Language;
+use function in_array;
 use Symfony\Component\Form\ChoiceList\ArrayChoiceList;
+use Symfony\Component\Form\ChoiceList\ChoiceListInterface;
 use Symfony\Component\Form\ChoiceList\Loader\ChoiceLoaderInterface;
 
 class ContentCreateLanguageChoiceLoader implements ChoiceLoaderInterface
 {
-    /** @var \Ibexa\AdminUi\Form\Type\ChoiceList\Loader\LanguageChoiceLoader */
-    private $languageChoiceLoader;
+    private LanguageChoiceLoader $languageChoiceLoader;
 
     /** @var string[] */
-    private $restrictedLanguagesCodes;
+    private array $restrictedLanguagesCodes;
 
     /**
-     * @param \Ibexa\AdminUi\Form\Type\ChoiceList\Loader\LanguageChoiceLoader $languageChoiceLoader
-     * @param array $restrictedLanguagesCodes
+     * @param array<string> $restrictedLanguagesCodes
      */
     public function __construct(
         LanguageChoiceLoader $languageChoiceLoader,
@@ -32,10 +32,7 @@ class ContentCreateLanguageChoiceLoader implements ChoiceLoaderInterface
         $this->restrictedLanguagesCodes = $restrictedLanguagesCodes;
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function loadChoiceList($value = null)
+    public function loadChoiceList(?callable $value = null): ChoiceListInterface
     {
         $languages = $this->languageChoiceLoader->getChoiceList();
 
@@ -44,16 +41,16 @@ class ContentCreateLanguageChoiceLoader implements ChoiceLoaderInterface
         }
 
         $languages = array_filter($languages, function (Language $language) {
-            return \in_array($language->languageCode, $this->restrictedLanguagesCodes, true);
+            return in_array($language->languageCode, $this->restrictedLanguagesCodes, true);
         });
 
         return new ArrayChoiceList($languages, $value);
     }
 
     /**
-     * {@inheritdoc}
+     * @return string[]
      */
-    public function loadChoicesForValues(array $values, $value = null)
+    public function loadChoicesForValues(array $values, ?callable $value = null): array
     {
         // Optimize
         $values = array_filter($values);
@@ -65,9 +62,9 @@ class ContentCreateLanguageChoiceLoader implements ChoiceLoaderInterface
     }
 
     /**
-     * {@inheritdoc}
+     * @return string[]
      */
-    public function loadValuesForChoices(array $choices, $value = null)
+    public function loadValuesForChoices(array $choices, ?callable $value = null): array
     {
         // Optimize
         $choices = array_filter($choices);
