@@ -12,16 +12,14 @@ use Ibexa\Contracts\Core\Repository\Values\Notification\NotificationList;
 use Pagerfanta\Adapter\AdapterInterface;
 
 /**
- * Pagerfanta adapter for Ibexa content search.
- * Will return results as notification list.
+ * @implements \Pagerfanta\Adapter\AdapterInterface<\Ibexa\Contracts\Core\Repository\Values\Notification\Notification>
  */
 class NotificationAdapter implements AdapterInterface
 {
-    /** @var \Ibexa\Contracts\Core\Repository\NotificationService */
-    private $notificationService;
+    private NotificationService $notificationService;
 
-    /** @var int */
-    private $nbResults;
+    /** @phpstan-var int<0, max> */
+    private int $nbResults;
 
     /**
      * @param \Ibexa\Contracts\Core\Repository\NotificationService $notificationService
@@ -32,33 +30,16 @@ class NotificationAdapter implements AdapterInterface
         $this->notificationService = $notificationService;
     }
 
-    /**
-     * Returns the number of results.
-     *
-     * @return int the number of results
-     */
     public function getNbResults(): int
     {
-        if ($this->nbResults !== null) {
-            return $this->nbResults;
-        }
-
-        return $this->nbResults = $this->notificationService->getNotificationCount();
+        return $this->nbResults ?? ($this->nbResults = $this->notificationService->getNotificationCount());
     }
 
-    /**
-     * Returns a slice of the results.
-     *
-     * @param int $offset the offset
-     * @param int $length the length
-     *
-     * @return \Ibexa\Contracts\Core\Repository\Values\Notification\NotificationList
-     */
-    public function getSlice($offset, $length): NotificationList
+    public function getSlice(int $offset, int $length): NotificationList
     {
         $notifications = $this->notificationService->loadNotifications($offset, $length);
 
-        if (null === $this->nbResults) {
+        if (!isset($this->nbResults)) {
             $this->nbResults = $notifications->totalCount;
         }
 
