@@ -11,17 +11,14 @@ use Ibexa\Contracts\Core\Repository\URLService;
 use Ibexa\Contracts\Core\Repository\Values\URL\URLQuery;
 use Pagerfanta\Adapter\AdapterInterface;
 
+/**
+ * @implements \Pagerfanta\Adapter\AdapterInterface<\Ibexa\Contracts\Core\Repository\Values\URL\URL>
+ */
 class URLSearchAdapter implements AdapterInterface
 {
-    /**
-     * @var \Ibexa\Contracts\Core\Repository\Values\URL\URLQuery
-     */
-    private $query;
+    private URLQuery $query;
 
-    /**
-     * @var \Ibexa\Contracts\Core\Repository\URLService
-     */
-    private $urlService;
+    private URLService $urlService;
 
     /**
      * UrlSearchAdapter constructor.
@@ -36,7 +33,7 @@ class URLSearchAdapter implements AdapterInterface
     }
 
     /**
-     * {@inheritdoc}
+     * @throws \Ibexa\Contracts\Core\Repository\Exceptions\UnauthorizedException
      */
     public function getNbResults(): int
     {
@@ -44,15 +41,16 @@ class URLSearchAdapter implements AdapterInterface
         $query->offset = 0;
         $query->limit = 0;
 
+        /** @phpstan-var int<0, max> */
         return $this->urlService->findUrls($query)->totalCount;
     }
 
     /**
-     * {@inheritdoc}
-     *
      * @return \Ibexa\Contracts\Core\Repository\Values\URL\URL[]
+     *
+     * @throws \Ibexa\Contracts\Core\Repository\Exceptions\UnauthorizedException
      */
-    public function getSlice($offset, $length): array
+    public function getSlice(int $offset, int $length): array
     {
         $query = clone $this->query;
         $query->offset = $offset;
