@@ -4,6 +4,7 @@
  * @copyright Copyright (C) Ibexa AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  */
+declare(strict_types=1);
 
 namespace Ibexa\Contracts\AdminUi\Controller;
 
@@ -13,17 +14,11 @@ use Symfony\Component\HttpFoundation\RedirectResponse;
 
 abstract class Controller extends AbstractController
 {
-    public function performAccessCheck()
+    public function performAccessCheck(): void
     {
         $this->denyAccessUnlessGranted('IS_AUTHENTICATED_REMEMBERED');
     }
 
-    /**
-     * @param \Ibexa\Contracts\Core\Repository\Values\Content\Location $location
-     * @param string $uriFragment
-     *
-     * @return \Symfony\Component\HttpFoundation\RedirectResponse
-     */
     public function redirectToLocation(Location $location, string $uriFragment = ''): RedirectResponse
     {
         return $this->redirectToRoute('ibexa.content.view', [
